requirements.txt
# Python3.12 or higher

requests
telethon
python-telegram-bot
